Move utilities to tool card location
